test: moved tests to folder
