+++ src/types/index.ts
@@ -1,5 +1,5 @@
 import type { Analytics } from "@ogcio/analytics-sdk";
-import type { FeatureFlags } from "../client/clients/featureFlags/index.js";
+import type { FeatureFlags } from "../client/clients/feature-flags/index.js";
 import type { Messaging } from "../client/clients/messaging/index.js";
 import type { Payments } from "../client/clients/payments/index.js";
 import type { Profile } from "../client/clients/profile/index.js";
